Move rune display functions to new file
