+++ src/bundles/rune/functions.ts
@@ -1,12 +1,9 @@
-import context from 'js-slang/context';
 import { mat4, vec3 } from 'gl-matrix';
 import {
   Rune,
-  NormalRune,
-  type RuneAnimation,
   DrawnRune,
   drawRunesToFrameBuffer,
-  AnimatedRune,
+  type AnimatedRune,
 } from './rune';
 import {
   getSquare,
@@ -31,11 +28,6 @@ import {
   initFramebufferObject,
   initShaderProgram,
 } from './runes_webgl';
-
-const drawnRunes: (DrawnRune | AnimatedRune)[] = [];
-context.moduleContexts.rune.state = {
-  drawnRunes,
-};
 
 export type RuneModuleState = {
   drawnRunes: (DrawnRune | AnimatedRune)[]
@@ -691,23 +683,6 @@ export function white(rune: Rune): Rune {
   return addColorFromHex(rune, '#FFFFFF');
 }
 
-// =============================================================================
-// Drawing functions
-// =============================================================================
-
-/**
- * Renders the specified Rune in a tab as a basic drawing.
- * @param rune - The Rune to render
- * @return {Rune} The specified Rune
- *
- * @category Main
- */
-export function show(rune: Rune): Rune {
-  throwIfNotRune(show.name, rune);
-  drawnRunes.push(new NormalRune(rune));
-  return rune;
-}
-
 /** @hidden */
 export class AnaglyphRune extends DrawnRune {
   private static readonly anaglyphVertexShader = `
@@ -816,20 +791,6 @@ export class AnaglyphRune extends DrawnRune {
     gl.enableVertexAttribArray(vertexPositionPointer);
     gl.drawArrays(gl.TRIANGLES, 0, 6);
   };
-}
-
-/**
- * Renders the specified Rune in a tab as an anaglyph. Use 3D glasses to view the
- * anaglyph.
- * @param rune - The Rune to render
- * @return {Rune} The specified Rune
- *
- * @category Main
- */
-export function anaglyph(rune: Rune): Rune {
-  throwIfNotRune(anaglyph.name, rune);
-  drawnRunes.push(new AnaglyphRune(rune));
-  return rune;
 }
 
 /** @hidden */
@@ -951,77 +912,3 @@ export class HollusionRune extends DrawnRune {
 
 /** @hidden */
 export const isHollusionRune = (rune: DrawnRune): rune is HollusionRune => rune.isHollusion;
-
-/**
- * Renders the specified Rune in a tab as a hollusion, using the specified
- * magnitude.
- * @param rune - The Rune to render
- * @param {number} magnitude - The hollusion's magnitude
- * @return {Rune} The specified Rune
- *
- * @category Main
- */
-export function hollusion_magnitude(rune: Rune, magnitude: number): Rune {
-  throwIfNotRune(hollusion_magnitude.name, rune);
-  drawnRunes.push(new HollusionRune(rune, magnitude));
-  return rune;
-}
-
-/**
- * Renders the specified Rune in a tab as a hollusion, with a default magnitude
- * of 0.1.
- * @param rune - The Rune to render
- * @return {Rune} The specified Rune
- *
- * @category Main
- */
-export function hollusion(rune: Rune): Rune {
-  throwIfNotRune(hollusion.name, rune);
-  return hollusion_magnitude(rune, 0.1);
-}
-
-/**
- * Create an animation of runes
- * @param duration Duration of the entire animation in seconds
- * @param fps Duration of each frame in frames per seconds
- * @param func Takes in the timestamp and returns a Rune to draw
- * @returns A rune animation
- *
- * @category Main
- */
-export function animate_rune(
-  duration: number,
-  fps: number,
-  func: RuneAnimation,
-) {
-  const anim = new AnimatedRune(duration, fps, (n) => {
-    const rune = func(n);
-    throwIfNotRune(animate_rune.name, rune);
-    return new NormalRune(rune);
-  });
-  drawnRunes.push(anim);
-  return anim;
-}
-
-/**
- * Create an animation of anaglyph runes
- * @param duration Duration of the entire animation in seconds
- * @param fps Duration of each frame in frames per seconds
- * @param func Takes in the timestamp and returns a Rune to draw
- * @returns A rune animation
- *
- * @category Main
- */
-export function animate_anaglyph(
-  duration: number,
-  fps: number,
-  func: RuneAnimation,
-) {
-  const anim = new AnimatedRune(duration, fps, (n) => {
-    const rune = func(n);
-    throwIfNotRune(animate_anaglyph.name, rune);
-    return new AnaglyphRune(rune);
-  });
-  drawnRunes.push(anim);
-  return anim;
-}
